+++ packages/kit/test/apps/basics/src/routes/routing/index.svelte
@@ -9,4 +9,6 @@
 <a href="http://localhost:{$page.url.searchParams.get('port')}">elsewhere</a>
 <a href="/static.json">static.json</a>
 
+<a href="/routing/b" sveltekit:reload>b</a>
+
 <div class="hydrate-test" />